requirements.txt
discord.py~=2.3.1
pymongo~=4.4.0
python-dotenv~=1.0.0
python-a2s~=1.3.0
